Fixed dyn_solver_t's tree caller
dyn_solver_t forms a tree, but there was an issue where multiple
nodes in the tree were referring to the same thing. This meant that
the load_balance went out of whack.

Looks like  it was mostly an
issue when calling dyn_solve_placement with 1 node--which has 1
trivial solution--and it mostly never happened.

Nevertheless, the solution is to make sure not to add duplicate
nodes.